clean up price service, tests
