+++ src/adapters/http/user.controller.ts
@@ -13,9 +13,8 @@ import {
     Res,
     UseGuards,
 } from "@nestjs/common";
-import { IsEmail, IsString } from "class-validator";
 import { Response } from "express";
-import { ActionStatus, BaseHttpDto } from "src/adapters/http/http.types";
+import { ActionStatus, BaseHttpDto, UpdateUserHttpDto, UserHttpDto } from "src/adapters/http/http.types";
 import { AuthServicePort } from "src/core/auth/api/auth.service.port";
 import { AuthToken } from "src/core/auth/api/auth.types";
 import { CreateUserDto, UpdateUserDto, UserDto } from "src/core/user/api/user.dto";
@@ -23,14 +22,6 @@ import { UserServicePort } from "src/core/user/api/user.service.port";
 import { AUTH_TOKEN_NAME, AuthenticatedRequest } from "./auth/auth.types";
 import { JwtAuthGuard } from "./auth/jwt.auth.guard";
 
-export class UpdateUserHttpDto {
-    @IsString() readonly name: string;
-    @IsEmail() readonly email: string;
-}
-
-export class UserHttpDto extends BaseHttpDto {
-    readonly user: UserDto;
-}
 
 @Controller("user")
 export class UserController {
