+++ src/core/price/api/price.service.port.ts
@@ -1,11 +1,10 @@
 import { CreatePriceDto } from "src/core/price/api/create-price.dto";
-import { PriceDto } from "src/core/price/api/price.dto";
 
 export const PriceServicePort = Symbol("PriceServicePort");
 
 /**
  * Interface representing the contract for a Price Service.
- * Provides methods for saving and retrieving price data associated with cards.
+ * Read ops handled by other services, this service is for write ops.
  */
 export interface PriceServicePort {
     /**
@@ -14,14 +13,6 @@ export interface PriceServicePort {
      * @param prices - An array of price data transfer objects to be saved.
      */
     save(price: CreatePriceDto[]): Promise<void>;
-
-    /**
-     * Finds a price entry by the card's unique identifier.
-     * 
-     * @param cardId - The unique identifier of the card.
-     * @returns A promise that resolves to the price data transfer object or null if not found.
-     */
-    findByCardId(cardId: number): Promise<PriceDto | null>;
 
     /**
      * Deletes a price entry by its unique identifier.
